+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - R32 en M5 F10 hebben opnieuw audio
 - Je ziet nu de prijs incl tax bij de voertuigrental
 - Je voertuig zal niet meer stallen door een meleehit
+- Fix Vehicles error wanneer dg-config gerestart
 
 ## [1.1.4]